Make states completely opaque
Removes `getState`, which allowed reading the `a` instead a `State s a`.
This allows state machines to keep their states completely private if
they like. You can recover the ability for your state to be read by
adding a transition using the `Const` functor. This relies on
transitions not linearly consuming their argument, so any future
hypothetical `LinearTypes` integration would need to rethink this.